Refactor billing system from 3-table architecture (accounts + billings +
allowances) to
State+Ledger pattern where accounts hold live state and billings serve
as an immutable
ledger with signed deltas and post-transaction snapshots. All monetary
values converted
from float to int64 micros (1 USD = 1,000,000) to eliminate
floating-point rounding errors.

- bin-billing-manager: Replace float balance with int64 balance_credit
and balance_token on Account
- bin-billing-manager: Add ledger fields to Billing (transaction_type,
amount_token, amount_credit, snapshots, idempotency_key)
- bin-billing-manager: Add CalculateBillableUnits for ceiling-rounded
per-minute billing
- bin-billing-manager: Add CalculateTokenCreditDeduction for token-first
credit-overflow logic
- bin-billing-manager: Add BillingConsumeAndRecord atomic transaction
(SELECT FOR UPDATE + deduct + insert)
- bin-billing-manager: Add AccountTopUpTokens with deterministic UUID v5
idempotency
- bin-billing-manager: Add monthly top-up cron with paginated account
processing
- bin-billing-manager: Add GetCostInfo and CostType for rate lookup by
call/sms/number type
- bin-billing-manager: Remove allowancehandler package and all
allowance-related code
- bin-billing-manager: Add 25 edge case tests for deduction logic and 15
tests for billable units
- bin-api-manager: Update billing account and billing endpoints for new
int64 fields
- bin-api-manager: Remove allowance endpoints (GET allowances, GET
allowance)
- bin-api-manager: Fix float32 truncation in balance endpoints using
math.Round
- bin-api-manager: Fix missing return after AbortWithStatus(400)
- bin-api-manager: Update RST documentation for State+Ledger
architecture
- bin-common-handler: Update requesthandler to remove allowance RPC
methods
- bin-openapi-manager: Update OpenAPI spec with new ledger fields,
remove allowance schema
- bin-dbscheme-manager: Add Alembic migration for schema changes with
data migration
- docs: Add design document and implementation plan

+++ bin-billing-manager/models/account/plan_test.go
@@ -139,7 +139,7 @@ func TestPlanTokenMap(t *testing.T) {
 	tests := []struct {
 		name     string
 		planType PlanType
-		expected int
+		expected int64
 	}{
 		{"free plan", PlanTypeFree, 1000},
 		{"basic plan", PlanTypeBasic, 10000},
